Update Product Controller

+++ backend/repositories/productRepository.js
@@ -1,23 +1,59 @@
 import Product from "../models/productModel.js";
 
 class ProductRepository {
+
     async createProduct(data) {
+        const { name, description, price } = data;
+        if (!name || !description || !price) {
+            throw new Error("Invalid Information");
+        }
         const existProduct = await Product.findOne({ name: data.name });
         if (existProduct) {
-            throw new Error("Product already exists")
+            throw new Error("Product already exists");
         }
-        return await Product.create(data);
+        const newProduct = await Product.create({ name, description, price });
+        if (!newProduct) {
+            throw new Error("Unable to create product");
+        } else {
+            return newProduct;
+        }
     };
+
     async getAllProduct() {
         return await Product.find({});
     };
+
     async getProductById(id) {
-        return await Product.findById(id);
+        if (!id) {
+            throw new Error("Invalid Information");
+        };
+        const foundProduct = await Product.findById(id);
+        if (!foundProduct) {
+            throw new Error("Product not found")
+        } else {
+            return foundProduct;
+        }
     };
+
     async updateProduct(id, data) {
-        return await Product.findByIdAndUpdate(id, data, { new: true });
+        const { name, description, price } = data;
+        if (!id || !name || !description || !price) {
+            throw new Error("Invalid Information");
+        };
+
+        const updateProduct = await Product.findByIdAndUpdate(id, data, { new: true });
+
+        if (!updateProduct) {
+            throw new Error("Unable to update the product");
+        } else {
+            return updateProduct;
+        }
     };
+
     async deleteProduct(id) {
+        if (!id) {
+            throw new Error("Invalid Information");
+        }
         return await Product.findByIdAndDelete(id);
     };
 };
